Move logic to fishRenderer
