+++ src/demo/constants.js
@@ -1,0 +1,6 @@
+const constants = {
+  canvasWidth: 1024,
+  canvasHeight: 576
+};
+
+export default constants;
